The auth token now only includes groups configured in Verdaccio.
Users can likely have access to dozens of GitHub repositories and teams.
Including all of these in the token is problematic because it quickly
increases the length of the token, eventually causing HTTP 431 errors.

To avoid this problem, auth tokens now only includes the required GitHub
org plus groups used in the Verdaccio config to configure package access.

Fixes #139